demux: libmp4: read the dec3 channel location in uint8_t
It is read with MP4_GET1BYTE(). We don't make use of the value anyway.

libavformat doesn't read this value.

+++ modules/demux/mp4/libmp4.h
@@ -1422,7 +1422,7 @@ typedef struct
         uint8_t i_acmod;
         uint8_t i_lfeon;
         uint8_t i_num_dep_sub;
-        uint16_t i_chan_loc;
+        uint8_t i_chan_loc;
     } stream[8];
 
 } MP4_Box_data_dec3_t;
